refactor(cli/test): move reporters to a separate directory (#19957)
Just a small cleanup that will make #19804 easier.
